# sofabaton-x-server/openapi.json
{
  "components": {
    "schemas": {
...
      "EntityIdentity": {
        "additionalProperties": true,
        "description": "The ``device`` block of an entity payload: id and name are typed, the rest rides along.",
        "properties": {
          "device_id": {
            "title": "Device Id",
            "type": "integer"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "device_id"
        ],
        "title": "EntityIdentity",
        "type": "object"
      },
...
      "JobView": {
        "description": "One job as the API shows it.",
        "properties": {
          "cancellable": {
            "title": "Cancellable",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Problem"
              },
              {
                "type": "null"
              }
            ]
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "hub_id": {
            "title": "Hub Id",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WriteProgress"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "done",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "hub_id",
          "kind",
          "status",
          "cancellable",
          "created_at"
        ],
        "title": "JobView",
        "type": "object"
      },
...
      "Problem": {
        "description": "Error body (RFC 9457 shape, minus the URI scheme).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "hub_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hub Id"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
      "RefreshRequest": {
        "description": "Body of ``POST /hubs/{id}/snapshot/refresh``: one entity, or neither for the whole hub.",
        "properties": {
          "activity_id": {
            "anyOf": [
              {
                "maximum": 255.0,
                "minimum": 101.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Id"
          },
          "device_id": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          }
        },
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "SnapshotDocument": {
        "additionalProperties": true,
        "description": "``GET /hubs/{id}/snapshot``: the structural configuration plus its header.\n\nThe body is the library's ``hub_bundle`` with the snapshot header\nmerged in; ``snapshot_id`` is also the response ``ETag``.",
        "properties": {
          "activities": {
            "items": {
              "$ref": "#/components/schemas/SnapshotEntityPayload"
            },
            "title": "Activities",
            "type": "array"
          },
          "captured_at": {
            "title": "Captured At",
            "type": "string"
          },
          "complete": {
            "title": "Complete",
            "type": "boolean"
          },
          "devices": {
            "items": {
              "$ref": "#/components/schemas/SnapshotEntityPayload"
            },
            "title": "Devices",
            "type": "array"
          },
          "engine_generation": {
            "title": "Engine Generation",
            "type": "integer"
          },
          "hub": {
            "additionalProperties": true,
            "title": "Hub",
            "type": "object"
          },
          "payload_profile": {
            "title": "Payload Profile",
            "type": "string"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          },
          "stale_risk": {
            "title": "Stale Risk",
            "type": "boolean"
          }
        },
        "required": [
          "snapshot_id",
          "captured_at",
          "engine_generation",
          "complete",
          "stale_risk",
          "payload_profile"
        ],
        "title": "SnapshotDocument",
        "type": "object"
      },
      "SnapshotEntityPayload": {
        "additionalProperties": true,
        "description": "One device or activity in the snapshot: provenance typed, tables open.\n\n``complete`` / ``editable`` / ``stale_risk`` / ``fetched_at`` are the\nlibrary's per-entity provenance; ``commands``, ``button_bindings``,\n``macros``, ``favorite_slots`` and the other tables follow the\nlibrary's ``hub_bundle`` shape and are passed through as they are.",
        "properties": {
          "complete": {
            "default": false,
            "title": "Complete",
            "type": "boolean"
          },
          "device": {
            "$ref": "#/components/schemas/EntityIdentity"
          },
          "editable": {
            "default": false,
            "title": "Editable",
            "type": "boolean"
          },
          "fetched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "stale_risk": {
            "default": false,
            "title": "Stale Risk",
            "type": "boolean"
          }
        },
        "required": [
          "kind",
          "device"
        ],
        "title": "SnapshotEntityPayload",
        "type": "object"
      },
...
      "WriteProgress": {
        "description": "One progress report from a long-running operation (refresh, sync).\n\n``phase`` is the operation's own phase word (``preparing``, ``device``,\n``activity``, ``stale_check``, ``writing``, ``finalizing`` ...);\n``entity_kind`` / ``entity_id`` name the entity being worked on when\nthere is one; ``step_kind`` is the sync step kind while writing.",
        "properties": {
          "completed_steps": {
            "title": "Completed Steps",
            "type": "integer"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "entity_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Kind"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "step_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Kind"
          },
          "total_steps": {
            "title": "Total Steps",
            "type": "integer"
          }
        },
        "required": [
          "phase",
          "message",
          "completed_steps",
          "total_steps"
        ],
        "title": "WriteProgress",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "REST + WebSocket over the sofabaton-x library for Sofabaton X1 / X1S / X2 hubs. LAN service; no authentication in v1.\n\nWebSocket: `/api/v1/events` (optional `?hub_id=` filter, repeatable) streams WsHello once, then WsHubEvent / WsServerEvent / WsJobEvent / WsDropped messages (see components).",
    "title": "sofabaton-x-server",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/hubs/{hub_id}/jobs": {
      "get": {
        "operationId": "listJobs",
        "parameters": [
          {
            "in": "path",
            "name": "hub_id",
            "required": true,
            "schema": {
              "title": "Hub Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobView"
                  },
                  "title": "Response Listjobs",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Recent jobs on this hub, newest first",
        "tags": [
          "snapshot"
        ]
      }
    },
    "/api/v1/hubs/{hub_id}/jobs/{job_id}": {
      "delete": {
        "operationId": "cancelJob",
        "parameters": [
          {
            "in": "path",
            "name": "hub_id",
            "required": true,
            "schema": {
              "title": "Hub Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Cancel a running job (where the operation allows it)",
        "tags": [
          "snapshot"
        ]
      },
      "get": {
        "operationId": "getJob",
        "parameters": [
          {
            "in": "path",
            "name": "hub_id",
            "required": true,
            "schema": {
              "title": "Hub Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "One job: status, progress, result or error",
        "tags": [
          "snapshot"
        ]
      }
    },
...
    "/api/v1/hubs/{hub_id}/snapshot": {
      "get": {
        "operationId": "getSnapshot",
        "parameters": [
          {
            "in": "path",
            "name": "hub_id",
            "required": true,
            "schema": {
              "title": "Hub Id",
              "type": "string"
            }
          },
          {
            "description": "an ETag from an earlier read",
            "in": "header",
            "name": "If-None-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "an ETag from an earlier read",
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified (If-None-Match matched the ETag)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "The hub's structural configuration from the cache (no hub traffic)",
        "tags": [
          "snapshot"
        ]
      }
    },
    "/api/v1/hubs/{hub_id}/snapshot/refresh": {
      "post": {
        "operationId": "refreshSnapshot",
        "parameters": [
          {
            "in": "path",
            "name": "hub_id",
            "required": true,
            "schema": {
              "title": "Hub Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RefreshRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Read structural detail from the hub as a job (one entity, or the whole hub)",
        "tags": [
          "snapshot"
        ]
      }
    },
...
  }
}